+++ shadcn_flutter/lib/shadcn_flutter.dart
@@ -68,7 +68,6 @@ export 'src/components/stage_container.dart';
 export 'src/components/steps.dart';
 export 'src/components/switch.dart';
 export 'src/components/tab_list.dart';
-export 'src/components/table.dart';
 export 'src/components/tabs.dart';
 export 'src/components/text/text.dart';
 export 'src/components/text/underline_text.dart';